2004-11-01	Joel Sherrill <joel@OARcorp.com>
	* score/include/rtems/score/object.h, score/src/object.c,
	score/src/objectallocatebyindex.c, score/src/objectcomparenameraw.c,
	score/src/objectcomparenamestring.c, score/src/objectcopynameraw.c,
	score/src/objectinitializeinformation.c: Fix bugs in previous commit.
	Now compiles.

+++ cpukit/score/include/rtems/score/object.h
@@ -240,8 +240,8 @@ typedef struct {
  */
 
 #if defined(RTEMS_MULTIPROCESSING)
-SCORE_EXTERN uint32_t    _Objects_Local_node;
-SCORE_EXTERN uint32_t    _Objects_Maximum_nodes;
+SCORE_EXTERN uint16_t    _Objects_Local_node;
+SCORE_EXTERN uint16_t    _Objects_Maximum_nodes;
 #else
 #define _Objects_Local_node    1
 #define _Objects_Maximum_nodes 1
